core/standards/historyinterface/007_14Mar2011.html

<!doctype html>
<html>
	<head>
		<title>Firing popstate after onload with pushed state</title>
		<script type="text/javascript" src="/resources/testharness.js"></script>
		<script type="text/javascript" src="/resources/testharnessreport.js"></script>
	</head>
	<body>

		<noscript><p>Enable JavaScript and reload</p></noscript>
		<div id="log">It looks like the browser stopped loading the page when encountering a .go(-1) command pointing to a pushed state. This will break the tests.</div>
		<script type="text/javascript">

//spec (14 March 2011 draft) states that popstate must fire after onload even if there is no pushed/replaced state
var popfired = false;
setup(function () {},{explicit_done:true,timeout:30000});
window.addEventListener('popstate',function (e) { popfired = e.state; },false);
test(function () {
	history.pushState('state1','');
	history.pushState('state2','');
	history.go(-1); //this causes Chrome to stop loading the test
}, 'history.pushState support is needed for this testcase');
window.onload = function () {
	test(function () {
		assert_false( popfired );
	}, 'popstate event should not fire before onload fires');
	popfired = false;
	setTimeout(function () {
		test(function () {
			assert_true( !!popfired );
		}, 'popstate event should fire after onload fires, even if there is no pushed/replaced state');
		test(function () {
			assert_equals( popfired, 'state1' );
		}, 'the correct state should be restored when navigating during initial load');
		done();
		if( history.pushState ) { history.go(-1); } //go back to the start to avoid state recovery when reloading
	},100);
};

		</script>

	</body>
</html>
